# internal/api/handlers/user_handler.go
package handlers

import (
	"go-backend-todo/internal/service"

	"github.com/gofiber/fiber/v2"
)

type UserHandler struct {
	userService service.UserService
}

func NewUserHandler(userService service.UserService) *UserHandler {
	return &UserHandler{
		userService: userService,
	}
}

// GetUser gets user by ID
// @Summary Get user by ID
// @Description Retrieve a user's information by their unique identifier
// @Tags Users
// @Accept json
// @Produce json
// @Param id path string true "User ID" format(uuid)
// @Security BearerAuth
// @Success 200 {object} models.UserAccount "User information"
// @Failure 400 {object} map[string]string "Invalid user ID format"
// @Failure 401 {object} map[string]string "Unauthorized - missing or invalid token"
// @Failure 404 {object} map[string]string "User not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users/{id} [get]
func (h *UserHandler) GetUser(c *fiber.Ctx) error {
	id := c.Params("id")

	return c.JSON(fiber.Map{
		"message": "Get user",
		"id":      id,
	})
}

// UpdateUser updates user
// @Summary Update user information
// @Description Update user's profile information such as username, email, or other details
// @Tags Users
// @Accept json
// @Produce json
// @Param id path string true "User ID" format(uuid)
// @Param user body models.UpdateUserRequest true "User update data"
// @Security BearerAuth
// @Success 200 {object} models.UserAccount "Updated user information"
// @Failure 400 {object} map[string]string "Invalid request data or user ID format"
// @Failure 401 {object} map[string]string "Unauthorized - missing or invalid token"
// @Failure 403 {object} map[string]string "Forbidden - cannot update other user's data"
// @Failure 404 {object} map[string]string "User not found"
// @Failure 409 {object} map[string]string "Conflict - username or email already exists"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users/{id} [put]
func (h *UserHandler) UpdateUser(c *fiber.Ctx) error {
	id := c.Params("id")
	// TODO: Implement logic to update user
	return c.JSON(fiber.Map{
		"message": "Update user",
		"id":      id,
	})
}

// DeleteUser deletes user
// @Summary Delete user account
// @Description Permanently delete a user account and all associated data
// @Tags Users
// @Accept json
// @Produce json
// @Param id path string true "User ID" format(uuid)
// @Security BearerAuth
// @Success 200 {object} map[string]string "User successfully deleted"
// @Failure 400 {object} map[string]string "Invalid user ID format"
// @Failure 401 {object} map[string]string "Unauthorized - missing or invalid token"
// @Failure 403 {object} map[string]string "Forbidden - cannot delete other user's account or admin required"
// @Failure 404 {object} map[string]string "User not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users/{id} [delete]
func (h *UserHandler) DeleteUser(c *fiber.Ctx) error {
	id := c.Params("id")
	// TODO: Implement logic to delete user
	return c.JSON(fiber.Map{
		"message": "Delete user",
		"id":      id,
	})
}

// GetAllUsers gets all users with pagination
// @Summary Get all users
// @Description Retrieve a paginated list of all users in the system
// @Tags Users
// @Accept json
// @Produce json
// @Param page query int false "Page number (default: 1)" minimum(1)
// @Param limit query int false "Items per page (default: 10)" minimum(1) maximum(100)
// @Param search query string false "Search users by username or email"
// @Security BearerAuth
// @Success 200 {object} map[string]interface{} "Paginated list of users"
// @Failure 400 {object} map[string]string "Invalid pagination parameters"
// @Failure 401 {object} map[string]string "Unauthorized - missing or invalid token"
// @Failure 403 {object} map[string]string "Forbidden - admin access required"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users [get]
func (h *UserHandler) GetAllUsers(c *fiber.Ctx) error {
	// TODO: Implement logic to get all users with pagination
	return c.JSON(fiber.Map{
		"message": "Get all users",
		"users":   []string{},
	})
}

// GetUserProfile gets current user's profile
// @Summary Get current user profile
// @Description Retrieve the profile information of the currently authenticated user
// @Tags Users
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} models.UserAccount "Current user profile"
// @Failure 401 {object} map[string]string "Unauthorized - missing or invalid token"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users/profile [get]
func (h *UserHandler) GetUserProfile(c *fiber.Ctx) error {
	// TODO: Implement logic to get current user profile
	return c.JSON(fiber.Map{
		"message": "Get user profile",
	})
}

// UpdateUserProfile updates current user's profile
// @Summary Update current user profile
// @Description Update the profile information of the currently authenticated user
// @Tags Users
// @Accept json
// @Produce json
// @Param profile body models.UpdateProfileRequest true "Profile update data"
// @Security BearerAuth
// @Success 200 {object} models.UserAccount "Updated user profile"
// @Failure 400 {object} map[string]string "Invalid request data"
// @Failure 401 {object} map[string]string "Unauthorized - missing or invalid token"
// @Failure 409 {object} map[string]string "Conflict - username or email already exists"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users/profile [put]
func (h *UserHandler) UpdateUserProfile(c *fiber.Ctx) error {
	// TODO: Implement logic to update current user profile
	return c.JSON(fiber.Map{
		"message": "Update user profile",
	})
}

// ChangePassword changes user's password
// @Summary Change user password
// @Description Change the password for the currently authenticated user
// @Tags Users
// @Accept json
// @Produce json
// @Param password body models.ChangePasswordRequest true "Password change data"
// @Security BearerAuth
// @Success 200 {object} map[string]string "Password changed successfully"
// @Failure 400 {object} map[string]string "Invalid request data or weak password"
// @Failure 401 {object} map[string]string "Unauthorized - missing or invalid token"
// @Failure 403 {object} map[string]string "Forbidden - current password is incorrect"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /users/change-password [put]
func (h *UserHandler) ChangePassword(c *fiber.Ctx) error {
	// TODO: Implement logic to change user password
	return c.JSON(fiber.Map{
		"message": "Change password",
	})
}